feat: rearranged the percussion so that the percussion objects take care of it rather than the PercussionManager
Gave each percussion object 4 audio sources to round robin between firing so that we can have more complicated echos and overlapping audio. There's still some clipping, but it kind of gives it a weird rhythmic effect, so I'm not going to worry about it too much right now. Also changed the audio clip to be more of a bell chime. Makes it a bit crisper? A little too on the nose, but will be good for the loud RadioSnack event